core: move NameResolverProvider to internal
