+++ core/src/main/java/io/grpc/internal/NameResolverProvider.java
@@ -14,10 +14,12 @@
  * limitations under the License.
  */
 
-package io.grpc;
+package io.grpc.internal;
 
 import com.google.common.annotations.VisibleForTesting;
 import com.google.common.base.Preconditions;
+import io.grpc.Attributes;
+import io.grpc.NameResolver;
 import java.net.URI;
 import java.util.ArrayList;
 import java.util.Collections;
@@ -33,7 +35,6 @@ import java.util.ServiceLoader;
  * exceptions may reasonably occur for implementation-specific reasons, implementations should
  * generally handle the exception gracefully and return {@code false} from {@link #isAvailable()}.
  */
-@Internal
 public abstract class NameResolverProvider extends NameResolver.Factory {
   /**
    * The port number used in case the target or the underlying naming system doesn't provide a
@@ -145,7 +146,7 @@ public abstract class NameResolverProvider extends NameResolver.Factory {
    * Whether this provider is available for use, taking the current environment into consideration.
    * If {@code false}, no other methods are safe to be called.
    */
-  protected abstract boolean isAvailable();
+  public abstract boolean isAvailable();
 
   /**
    * A priority, from 0 to 10 that this provider should be used, taking the current environment into
@@ -153,7 +154,7 @@ public abstract class NameResolverProvider extends NameResolver.Factory {
    * detection. A priority of 0 does not imply that the provider wouldn't work; just that it should
    * be last in line.
    */
-  protected abstract int priority();
+  public abstract int priority();
 
   private static class NameResolverFactory extends NameResolver.Factory {
     private final List<NameResolverProvider> providers;
