+++ core/src/main/resources/META-INF/services/io.grpc.internal.NameResolverProvider
