+++ core/src/test/java/io/grpc/internal/NameResolverProviderTest.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package io.grpc;
+package io.grpc.internal;
 
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertFalse;
@@ -24,7 +24,9 @@ import static org.junit.Assert.assertTrue;
 import static org.junit.Assert.fail;
 import static org.mockito.Mockito.mock;
 
-import io.grpc.internal.DnsNameResolverProvider;
+import io.grpc.Attributes;
+import io.grpc.NameResolver;
+import io.grpc.ReplacingClassLoader;
 import java.net.URI;
 import java.util.Collections;
 import java.util.List;
@@ -36,7 +38,7 @@ import org.junit.runners.JUnit4;
 /** Unit tests for {@link NameResolverProvider}. */
 @RunWith(JUnit4.class)
 public class NameResolverProviderTest {
-  private final String serviceFile = "META-INF/services/io.grpc.NameResolverProvider";
+  private final String serviceFile = "META-INF/services/io.grpc.internal.NameResolverProvider";
   private final URI uri = URI.create("dns:///localhost");
   private final Attributes attributes = Attributes.EMPTY;
 
@@ -44,7 +46,7 @@ public class NameResolverProviderTest {
   public void noProvider() {
     ClassLoader cl = new ReplacingClassLoader(
         getClass().getClassLoader(), serviceFile,
-        "io/grpc/NameResolverProviderTest-doesNotExist.txt");
+        "io/grpc/internal/NameResolverProviderTest-doesNotExist.txt");
     List<NameResolverProvider> providers = NameResolverProvider.load(cl);
     assertEquals(Collections.<NameResolverProvider>emptyList(), providers);
   }
@@ -53,7 +55,7 @@ public class NameResolverProviderTest {
   public void multipleProvider() {
     ClassLoader cl = new ReplacingClassLoader(
         getClass().getClassLoader(), serviceFile,
-        "io/grpc/NameResolverProviderTest-multipleProvider.txt");
+        "io/grpc/internal/NameResolverProviderTest-multipleProvider.txt");
     List<NameResolverProvider> providers = NameResolverProvider.load(cl);
     assertEquals(3, providers.size());
     assertSame(Available7Provider.class, providers.get(0).getClass());
@@ -69,7 +71,7 @@ public class NameResolverProviderTest {
   public void unavailableProvider() {
     ClassLoader cl = new ReplacingClassLoader(
         getClass().getClassLoader(), serviceFile,
-        "io/grpc/NameResolverProviderTest-unavailableProvider.txt");
+        "io/grpc/internal/NameResolverProviderTest-unavailableProvider.txt");
     assertEquals(Collections.<NameResolverProvider>emptyList(), NameResolverProvider.load(cl));
   }
 
@@ -170,12 +172,12 @@ public class NameResolverProviderTest {
     }
 
     @Override
-    protected boolean isAvailable() {
+    public boolean isAvailable() {
       return isAvailable;
     }
 
     @Override
-    protected int priority() {
+    public int priority() {
       return priority;
     }
 
@@ -221,7 +223,7 @@ public class NameResolverProviderTest {
     }
 
     @Override
-    protected int priority() {
+    public int priority() {
       throw new RuntimeException("purposefully broken");
     }
   }
