--- core/src/test/resources/io/grpc/NameResolverProviderTest-multipleProvider.txt
@@ -1,3 +1,0 @@
-io.grpc.NameResolverProviderTest$Available5Provider
-io.grpc.NameResolverProviderTest$Available7Provider
-io.grpc.NameResolverProviderTest$Available0Provider
