--- core/src/test/resources/io/grpc/NameResolverProviderTest-unavailableProvider.txt
@@ -1,1 +1,0 @@
-io.grpc.NameResolverProviderTest$UnavailableProvider
