+++ core/src/test/resources/io/grpc/internal/NameResolverProviderTest-unavailableProvider.txt
@@ -1,0 +1,1 @@
+io.grpc.internal.NameResolverProviderTest$UnavailableProvider
